# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Stream Lights personal lighting control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /triggers:
    get:
      operationId: listTriggers
      tags: [triggers]
      summary: List all event triggers
      responses:
        "200":
          description: Triggers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trigger"
    post:
      operationId: createTrigger
      tags: [triggers]
      summary: Create an event trigger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerInput"
      responses:
        "201":
          description: Trigger created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trigger"
  /triggers/{id}:
    get:
      operationId: getTrigger
      tags: [triggers]
      summary: Get a trigger by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trigger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trigger"
    patch:
      operationId: updateTrigger
      tags: [triggers]
      summary: Update a trigger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trigger"
    delete:
      operationId: deleteTrigger
      tags: [triggers]
      summary: Delete a trigger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /triggers/{id}/fire:
    post:
      operationId: fireTrigger
      tags: [triggers]
      summary: Manually fire a trigger for testing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
components:
  schemas:
...
    Trigger:
      type: object
      required: [id, eventType, name, enabled]
      properties:
        id:
          type: integer
        name:
          type: string
        eventType:
          type: string
          description: "follow | subscribe | bits | donation | raid | host | ban | timeout | chat_message | channel_point"
        platform:
          type: ["string", "null"]
        enabled:
          type: boolean
        color:
          type: string
          description: Hex color like #FF0000
        brightness:
          type: integer
        durationMs:
          type: integer
        effect:
          type: string
          description: "solid | strobe | pulse | rainbow | fade | police"
        returnToIdle:
          type: boolean
        minAmount:
          type: ["integer", "null"]
          description: Minimum bits/donation to trigger
        deviceIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
    TriggerInput:
      type: object
      required: [name, eventType, color, brightness, durationMs, effect]
      properties:
        name:
          type: string
        eventType:
          type: string
        platform:
          type: string
        enabled:
          type: boolean
        color:
          type: string
        brightness:
          type: integer
        durationMs:
          type: integer
        effect:
          type: string
        returnToIdle:
          type: boolean
        minAmount:
          type: integer
        deviceIds:
          type: array
          items:
            type: integer
    TriggerUpdate:
      type: object
      properties:
        name:
          type: string
        platform:
          type: string
        enabled:
          type: boolean
        color:
          type: string
        brightness:
          type: integer
        durationMs:
          type: integer
        effect:
          type: string
        returnToIdle:
          type: boolean
        minAmount:
          type: integer
        deviceIds:
          type: array
          items:
            type: integer
...
    ActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...