# lib/api-zod/src/generated/types/trigger.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stream Lights personal lighting control API
 * OpenAPI spec version: 0.1.0
 */

export interface Trigger {
  id: number;
  name: string;
  /** follow | subscribe | bits | donation | raid | host | ban | timeout | chat_message | channel_point */
  eventType: string;
  /** @nullable */
  platform?: string | null;
  enabled: boolean;
  /** Hex color like */
  color?: string;
  brightness?: number;
  durationMs?: number;
  /** solid | strobe | pulse | rainbow | fade | police */
  effect?: string;
  returnToIdle?: boolean;
  /**
     * Minimum bits/donation to trigger
     * @nullable
     */
  minAmount?: number | null;
  deviceIds?: number[];
  createdAt?: string;
}