# lib/api-zod/src/generated/types/triggerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stream Lights personal lighting control API
 * OpenAPI spec version: 0.1.0
 */

export interface TriggerInput {
  name: string;
  eventType: string;
  platform?: string;
  enabled?: boolean;
  color: string;
  brightness: number;
  durationMs: number;
  effect: string;
  returnToIdle?: boolean;
  minAmount?: number;
  deviceIds?: number[];
}